1825 - Moved s3 related functions to utils/s3_file_utils.py
