+++ projects/_01_ingest/cqc_pir/jobs/ingest_cqc_pir_data.py
@@ -4,26 +4,27 @@ import sys
 
 os.environ["SPARK_VERSION"] = "3.5"
 
-from polars_utils.utils import (
+from schemas.cqc_pir_schema import PIR_SCHEMA
+from utils import utils
+from utils.s3_file_utils import (
     construct_destination_path,
     construct_s3_uri,
     get_s3_objects_list,
     is_csv,
+    split_s3_uri,
 )
-from schemas.cqc_pir_schema import PIR_SCHEMA
-from utils import utils
 
 
 def main(source, destination):
     if is_csv(source):
         print("Single file provided to job. Handling single file.")
-        bucket, key = utils.split_s3_uri(source)
+        bucket, key = split_s3_uri(source)
         new_destination = construct_destination_path(destination, key)
         ingest_pir_dataset(source, new_destination, PIR_SCHEMA)
         return
 
     print("Multiple files provided to job. Handling each file...")
-    bucket, prefix = utils.split_s3_uri(source)
+    bucket, prefix = split_s3_uri(source)
     objects_list = get_s3_objects_list(bucket, prefix)
 
     print("Objects list:")
